Pedantry! Pedantry everywhere! (All compiles with -Wall -Werror -ansi -pedantic now)
--HG--
branch : munge

+++ gelfshims/linux-gnu-libc6/readdir64.c
@@ -1,5 +1,6 @@
 #include <dirent.h>
 #include <stdio.h>
+#include <stdint.h>
 #include <stdlib.h>
 #include "shim.h"
 
@@ -7,8 +8,8 @@
 #define TREADDIR_C
 
 TSHIM(structdirent64) {
-    unsigned long long d_ino;
-    long long d_off;
+    uint64_t d_ino;
+    int64_t d_off;
     unsigned short int d_reclen;
     unsigned char d_type;
     char d_name[256];
